requirements.dash.2.12.1.txt
# python==3.9
scikit-learn
scikit-learn-intelex
pandas
numpy
scipy
joblib
pyyaml
tqdm
openpyxl
pillow
dash==2.12.1
flask==2.2.5
# diskcache
# celery
# redis
dash_bootstrap_components
dash-daq
chardet==4.0.0
transformers==4.26.1
accelerate
optimum==1.8.6
datasets
fugashi[unidic-lite]
ipadic
# ipython
# ipykernel
# notebook

## install PyToch
## https://pytorch.org/get-started/previous-versions/#v200
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.0+cpu
torchvision==0.15.1+cpu
torchaudio==2.0.1